updated return type

+++ tech/hanfeng/algorithms/array/MaximumAverageSubarray.kt
@@ -3,8 +3,8 @@ package tech.hanfeng.algorithms.array
 import java.lang.IllegalArgumentException
 import kotlin.math.max
 
-fun maxAverageSubarray(A : IntArray, k : Int) : Float {
-    if (A.isEmpty()) return 0f
+fun maxAverageSubarray(A : IntArray, k : Int) : Double {
+    if (A.isEmpty()) return 0.0
     if (k > A.size) throw IllegalArgumentException("k MUST be smaller than the given array size")
     if (k <= 0 || k > 30000) throw IllegalArgumentException("k MUST be greater than 0 and less than 30,000")
     if (A.size < -10000 || A.size > 10000) throw IllegalArgumentException("Array size MUST be in range of -10000 and 10000")
@@ -19,7 +19,7 @@ fun maxAverageSubarray(A : IntArray, k : Int) : Float {
         maxSum = max(maxSum, tmpSum)
     }
 
-    return maxSum.toFloat() / k
+    return maxSum.toDouble() / k
 }
 
 fun main() {
